% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ages.R
\name{group_ages}
\alias{group_ages}
\title{Group ages into categorical age groups}
\usage{
group_ages(ages, by = 1, max_age = 100)
}
\arguments{
\item{by}{age interval breaks}

\item{max_age}{maximum age after which ages are grouped into a max_age+ category e.g 100+}

\item{x}{a numeric vector}
}
\value{
vector of class factor
}
\description{
Group numeric ages into a factor using base::cut
}
\examples{
group_ages(ages = 1:100, by=5, max_age=95)

}
