% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document_moh.R
\name{html_document_moh}
\alias{html_document_moh}
\title{Convert to an MoH HTML document}
\usage{
html_document_moh(...)
}
\arguments{
\item{...}{passed onto \code {rmarkdown::html_document}}
}
\description{
Format for converting from R Markdown to an HTML document.
}
