% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit2prob.R
\name{logit2prob}
\alias{logit2prob}
\title{Logit to probability transformer}
\usage{
logit2prob(logit)
}
\arguments{
\item{logit}{numeric value or vector of logits}
}
\description{
Transform a logit to a probability
}
