% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moh_colours.R
\name{moh_colours}
\alias{moh_colours}
\title{NSW goverment branding colours}
\usage{
moh_colours(..., show_colours = FALSE)
}
\arguments{
\item{...}{Integer or characters vector. See details}

\item{show_colours}{Boolean. If TRUE a barplot of the colours with their names is shown}
}
\description{
A number of colours from https://www.digital.nsw.gov.au/design-system/style/colour
}
\details{
You can use indicies to select colours, or the colour names themselves. Use moh_colours(show_colours = T)
to see what is available
}
\examples{
# show colours and names
moh_colours(show_colours = TRUE)

# Return all colours
moh_colours() # all ten colours

# Return first 4 colours
moh_colours(1:4) # first 4 colours

# Return colours by name
moh_colours(c('darkBlue', 'red')) # colours by name ()

}
