% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moh_colours.R
\name{moh_ggplot}
\alias{moh_ggplot}
\alias{scale_fill_moh}
\alias{scale_colour_moh}
\title{NSW goverment branding colour scale for ggplot2}
\usage{
scale_fill_moh()

scale_colour_moh()
}
\description{
A number of colours from https://www.digital.nsw.gov.au/design-system/style/colour
}
\section{Functions}{
\itemize{
\item \code{scale_fill_moh()}: Scale fill

\item \code{scale_colour_moh()}: Scale color

}}
\examples{

require(ggplot2)

ggplot(
 data.frame(
   names = c('darkBlue', 'red', 'lightBlue',
             'medBlue', 'lightGrey', 'darkGrey',
             'black', 'yellow', 'orange', 'purple'),
   vals = rep(1,10)),
 aes(x = names, y = vals, fill = names)) +
 geom_col() +
 scale_fill_moh() +
 theme_void()

ggplot(
 data.frame(
   names = c('darkBlue', 'red', 'lightBlue',
             'medBlue', 'lightGrey', 'darkGrey',
             'black', 'yellow', 'orange', 'purple'),
   vals = rep(1,10)),
 aes(x = names, y = vals, colour = names)) +
 geom_col(size = 3, fill = 'white') +
 scale_colour_moh() +
 theme_void()


}
