% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit2prob.R
\name{prob2logit}
\alias{prob2logit}
\title{Probability to logit transformer}
\usage{
prob2logit(prob)
}
\arguments{
\item{prob}{numeric value or vector of probabilities between 0 and 1}
}
\description{
Transform a probability to a logit
}
